removed old vsearch file, reformatted some stuff
